Create types for optional dependencies in 'common' (#39)
This makes the 'cli' packages buildable without first building the optional packages.
