+++ packages/common/src/index.ts
@@ -1,2 +1,19 @@
 export { getMeticulousLocalDataDir } from "./local-data/local-data";
 export { DebugLogger } from "./logger/debug-logger";
+export type {
+  RecordSessionFn,
+  RecordSessionOptions,
+} from "./types/record.types";
+export type {
+  CreateReplayDebuggerFn,
+  ReplayDebuggerDependencies,
+  ReplayDebuggerOptions,
+} from "./types/replay-debugger.types";
+export type {
+  BaseReplayEventsDependencies,
+  ReplayEventsDependencies,
+  ReplayEventsDependency,
+  ReplayEventsFn,
+  ReplayEventsOptions,
+} from "./types/replay.types";
+export type { RecordedSession, SessionData } from "./types/session.types";
