+++ packages/common/src/types/record.types.ts
@@ -1,0 +1,20 @@
+import type { DebugLogger } from "../logger/debug-logger";
+
+export interface RecordSessionOptions {
+  browser: any;
+  project: any;
+  recordingToken: string;
+  appCommitHash: string;
+  devTools?: boolean | null | undefined;
+  verbose?: boolean | null | undefined;
+  recordingSnippet: string;
+  width?: number | null | undefined;
+  height?: number | null | undefined;
+  uploadIntervalMs?: number | null | undefined;
+  incognito?: boolean | null | undefined;
+  cookieDir?: string | null | undefined;
+  logger?: DebugLogger | null | undefined;
+  onDetectedSession?: (sessionId: string) => void;
+}
+
+export type RecordSessionFn = (options: RecordSessionOptions) => Promise<void>;
